# perf: various performance enhancements