--- a/e2e/flags/.snapshots/TestInitCommand
+++ b/e2e/flags/.snapshots/TestInitCommand
@@ -13,12 +13,12 @@ rule:
 scan:
     context: ""
     data_subject_mapping: ""
-    debug: false
     disable-domain-resolution: true
     domain-resolution-timeout: 3s
     external-rule-dir: []
     force: false
     internal-domains: []
+    log-level: info
     parallel: 0
     quiet: false
     scanner: