--- a/new/detector/composition/java/java.go
+++ b/new/detector/composition/java/java.go
@@ -10,11 +10,13 @@ import (
 	"github.com/bearer/bearer/pkg/commands/process/settings"
 	"github.com/bearer/bearer/pkg/report/customdetectors"
 	"github.com/bearer/bearer/pkg/util/file"
+	"github.com/rs/zerolog/log"
 
 	"github.com/bearer/bearer/new/detector/composition/types"
 	"github.com/bearer/bearer/new/detector/detection"
 	"github.com/bearer/bearer/new/detector/evaluator"
 	cachepkg "github.com/bearer/bearer/new/detector/evaluator/cache"
+	"github.com/bearer/bearer/new/detector/evaluator/stats"
 	"github.com/bearer/bearer/new/detector/implementation/custom"
 	"github.com/bearer/bearer/new/detector/implementation/generic/datatype"
 	"github.com/bearer/bearer/new/detector/implementation/generic/insecureurl"
@@ -38,9 +40,14 @@ type Composition struct {
 	lang                languagetypes.Language
 	closers             []func()
 	rules               map[string]*settings.Rule
+	stats               *stats.Stats
 }
 
-func New(rules map[string]*settings.Rule, classifier *classification.Classifier) (detectortypes.Composition, error) {
+func New(
+	debugProfile bool,
+	rules map[string]*settings.Rule,
+	classifier *classification.Classifier,
+) (detectortypes.Composition, error) {
 	lang, err := language.Get("java")
 	if err != nil {
 		return nil, fmt.Errorf("failed to lookup language: %s", err)
@@ -49,6 +56,10 @@ func New(rules map[string]*settings.Rule, classifier *classification.Classifier)
 	composition := &Composition{
 		langImplementation: java.Get(),
 		lang:               lang,
+	}
+
+	if debugProfile {
+		composition.stats = stats.New()
 	}
 
 	staticDetectors := []struct {
@@ -166,6 +177,10 @@ func New(rules map[string]*settings.Rule, classifier *classification.Classifier)
 }
 
 func (composition *Composition) Close() {
+	if composition.stats != nil {
+		log.Debug().Msgf("java stats:\n%s", composition.stats)
+	}
+
 	for _, closeFunc := range composition.closers {
 		closeFunc()
 	}
@@ -203,6 +218,7 @@ func (composition *Composition) DetectFromFileWithTypes(
 		composition.detectorSet,
 		tree,
 		file.FileInfo.Name(),
+		composition.stats,
 	)
 
 	sharedCache := cachepkg.NewShared(sharedDetectorTypes)