--- a/new/detector/detection/detection.go
+++ b/new/detector/detection/detection.go
@@ -1,0 +1,9 @@
+package detection
+
+import "github.com/bearer/bearer/new/language/tree"
+
+type Detection struct {
+	DetectorType string
+	MatchNode    *tree.Node
+	Data         interface{}
+}