… new/detector/composition/testhelper/testhelper.go
… new/detector/evaluator/cache/cache.go
… new/detector/evaluator/evaluator.go
… new/detector/evaluator/stats/stats.go
… new/detector/implementation/custom/custom.go
… new/detector/implementation/custom/filter.go
… new/detector/implementation/generic/datatype/datatype.go
… new/detector/implementation/generic/generic.go
… new/detector/implementation/generic/insecureurl/insecureurl.go
… new/detector/implementation/generic/stringliteral/stringliteral.go
… new/detector/implementation/generic/types/types.go
… new/detector/implementation/java/object/object.go
… new/detector/implementation/java/object/projection.go
… new/detector/implementation/java/string/string.go
… new/detector/implementation/javascript/object/object.go
… new/detector/implementation/javascript/object/projection.go
… new/detector/implementation/javascript/string/string.go
… new/detector/implementation/ruby/object/object.go
… new/detector/implementation/ruby/object/projection.go
… new/detector/implementation/ruby/string/string.go
… new/detector/implementation/testhelper/testhelper.go
… new/detector/set/set.go
… new/detector/types/types.go
… new/language/base/base.go
… new/language/implementation/implementation.go
… new/language/implementation/java/java.go
… new/language/implementation/javascript/javascript.go
… new/language/implementation/ruby/ruby.go
… new/language/patternquery/builder/builder.go
… new/language/patternquery/patternquery.go