## Description
<!-- What does this PR do and how does it -->

Improves performance by:
- Caching detections at the evaluation level rather than just the node level. This avoids repeated walking of the same part of the tree
- When a file times out, try to gracefully stop the scan in the worker rather than having to kill the process. This avoids reloading the rules in this case.
- Don't create a tree sitter query for patterns that are just a single variable. These match any node anyway.
- Sort filters in order of likely complexity before evaluation. This helps to throw away invalid matches before running the more expensive filters.

Benchmark:

|Project|Before|After|Speedup|
|-|-|-|-|
|Rocket Chat|2min 11sec|1min 36sec|1.4x|
|Juice Shop|1min 27sec|1min 13sec|1.2x|

_(numbers are from the file progress bar, not entire program execution time)_

Also:
- Fixes an issue where a debug profile would not be written if the worker was interrupted
- Adds `--log-level` option and adds some trace logging for rule evaluation.
- Adds rule timing to debug output when `--debug-profile` is set
- Changes log messages to pure text rather than JSON. This is much better when there are newlines in the output.
- Fixes a panic in the typescript global data types detector.

## Related
- Close #1065
<!-- References some existing PR
- #CCC
-->

## Checklist

- [ ] I've added test coverage that shows my fix or feature works as expected.
- [ ] I've updated or added documentation if required.
- [ ] I've included usage information in the description if CLI behavior was updated or added.
- [x] PR title follows [Conventional Commits](https://www.conventionalcommits.org/) format